backends/emu/requirements.txt
mypy==0.910
pytest
black==22.6.0
Flask==2.1.0
gunicorn==20.1.0
requests
jsonpath-ng
flask-cors
pysondb-v2